Move PLT account data to GlobalState/Account.hs
