+++ concordium-consensus/src/Concordium/GlobalState/Persistent/BlockState/ProtocolLevelTokens.hs
@@ -457,83 +457,3 @@ migrateProtocolLevelTokensForPV migration oldPLTsPV@(ProtocolLevelTokensForPV (C
         SFalse ->
             -- There are no migrations that remove PLTs altogether.
             case migration of {}
-
--- | The table of PLTs account states. The table is indexed by the token index
---  into the global token table.
-newtype TokenAccountStateTable = TokenAccountStateTable
-    { tokenAccountStateTable :: Map.Map TokenIndex (HashedBufferedRef TokenAccountState)
-    }
-    deriving newtype (Show)
-
-instance (MonadBlobStore m) => MHashableTo m SHA256.Hash TokenAccountStateTable where
-    getHashM (TokenAccountStateTable tast) = do
-        bs <-
-            mapM
-                ( \(tIx, ref) -> do
-                    h <- getHashM ref
-                    return (SHA256.hashOfHashes (getHash tIx) h)
-                )
-                $ Map.toAscList tast
-        return $ SHA256.hashLazy $ runPutLazy $ put bs
-
-instance (MonadBlobStore m) => BlobStorable m TokenAccountStateTable where
-    load = do
-        l <- getLength
-        m <- go l (return [])
-        return $ TokenAccountStateTable . Map.fromAscList <$> m
-      where
-        go 0 accM = return accM
-        go s accM = do
-            tIx <- get
-            mRef <- load
-            let accM' = do
-                    xs <- accM
-                    ref <- mRef
-                    return $ (tIx, ref) : xs
-            go (s - 1) accM'
-
-    storeUpdate TokenAccountStateTable{..} = do
-        let ss = Map.toAscList tokenAccountStateTable
-        let pLength = putLength $ length ss
-        (p, xs) <- go ss (pLength, [])
-        return (p, TokenAccountStateTable $ Map.fromAscList $ reverse xs)
-      where
-        go [] acc = return acc
-        go ((tIx, stateRef) : ys) (pAcc, xs) = do
-            (pRef, newRef) <- storeUpdate stateRef
-            go ys (pAcc >> put tIx >> pRef, (tIx, newRef) : xs)
-
--- | The empty token account state table.
-emptyTokenAccountStateTable :: TokenAccountStateTable
-emptyTokenAccountStateTable = TokenAccountStateTable{tokenAccountStateTable = Map.empty}
-
--- | The type of keys in the token state key-value map.
-type TokenAccountStateKey = SBS.ShortByteString
-
--- | The type of values in the token state key-value map.
-type TokenAccountStateValue = BS.ByteString
-
--- | Token state at the account level
-data TokenAccountState = TokenAccountState
-    { -- | The available balance for the account.
-      tasBalance :: !TokenRawAmount,
-      -- | The token module state for the account, represented as a key-value map.
-      tasModuleState :: !(Map.Map TokenAccountStateKey TokenAccountStateValue)
-    }
-    deriving (Eq, Show, Ord)
-
-instance Serialize TokenAccountState where
-    put TokenAccountState{..} = do
-        put tasBalance
-        put tasModuleState
-    get = do
-        tasBalance <- get
-        tasModuleState <- get
-        return TokenAccountState{..}
-
-instance HashableTo SHA256.Hash TokenAccountState where
-    getHash = SHA256.hashLazy . runPutLazy . put
-
-instance (Monad m) => MHashableTo m SHA256.Hash TokenAccountState
-
-instance (MonadBlobStore m) => BlobStorable m TokenAccountState
